FossilOrigin-Name: af7a8d50f3902c26edff6a76f36eb84532b9fef59774e5a9591ab94ad56a9edc

+++ src/hctInt.h
@@ -54,7 +54,7 @@ int sqlite3HctTreeBegin(HctTree *pTree, int iStmt);
 int sqlite3HctTreeRelease(HctTree *pTree, int iStmt);
 int sqlite3HctTreeRollbackTo(HctTree *pTree, int iStmt);
 
-int sqlite3HctTreeClearOne(HctTree *pTree, u32 iRoot, int *pnRow);
+int sqlite3HctTreeClearOne(HctTree *pTree, u32 iRoot, i64 *pnRow);
 
 int sqlite3HctTreeCsrOpen(HctTree *pTree, u32 iRoot, HctTreeCsr **ppCsr);
 int sqlite3HctTreeCsrClose(HctTreeCsr *pCsr);
